fix(emit): no spurious indent before semicolon after missing binary RHS
When source ends with a trailing newline after a binary `=` operator and the
RHS is a missing expression (parse-error recovery, TS1109), the emitter
detected `has_newline_after_op=true` and called `increase_indent()` before
emitting the absent RHS. The empty-string write inside `emit_identifier` then
triggered `ensure_indent()` which materialized 4 spaces into the output,
causing the statement-level `;` to be incorrectly indented.

Fix: skip `increase_indent`/`decrease_indent` in the `has_newline_after_op`
path when the RHS is a missing recovery identifier. This matches tsc's output
for `parserPrivateIdentifierInArrayAssignment` baseline.

Structural rule: a missing/zero-width RHS carries no content that warrants
continuation indentation; the statement-level semicolon must land at the
ambient indent of the enclosing statement.

Also split `tests_navigation.rs` (2044 lines) into `tests_navigation.rs` +
`tests_navigation_2.rs` to satisfy the §19 2000-line hard limit.
